added flags for green yellow red lights and controller type

+++ lights/controller/gpio_controller.go
@@ -12,7 +12,7 @@ type GpioController struct {
 	red    rpio.Pin
 }
 
-func NewGpioController(greenPinId int8, yellowPinId int8, redPinId int8) *GpioController {
+func NewGpioController(greenPinId int, yellowPinId int, redPinId int) *GpioController {
 	err := rpio.Open()
 	if err != nil {
 		log.Println("Cannot allocate RAM for GPIOs.")
